+++ src/data/adHocFilter.ts
@@ -1,3 +1,5 @@
+import sqlToAST, { clausesToSql, Clause } from './ast';
+
 export class AdHocFilter {
   private _targetTable = '';
 
@@ -23,48 +25,9 @@ export class AdHocFilter {
     }
     // Semicolons are not required and cause problems when building the SQL
     sql = sql.replace(';', '');
-    const ast = this.sqlToAST(sql);
+    const ast = sqlToAST(sql);
     const filteredAST = this.applyFiltersToAST(ast, whereClause);
-    return this.clausesToSql(filteredAST);
-  }
-
-  private sqlToAST(sql: string): Map<string, Clause> {
-    const ast = this.createStatement();
-    const re =
-      /\b(WITH|SELECT|DISTINCT|FROM|SAMPLE|JOIN|PREWHERE|WHERE|GROUP BY|LIMIT BY|HAVING|LIMIT|OFFSET|UNION|INTERSECT|EXCEPT|INTO OUTFILE|FORMAT)\b/gi;
-    let bracketCount = 0;
-    let lastBracketCount = 0;
-    let lastNode = '';
-    let bracketPhrase = '';
-    let regExpArray: RegExpExecArray | null;
-    while ((regExpArray = re.exec(sql)) !== null) {
-      // Sets foundNode to a SQL keyword from the regular expression
-      const foundNode = regExpArray[0].toUpperCase();
-      const phrase = sql.substring(re.lastIndex, sql.length).split(re)[0];
-      // If there is a greater number of open brackets than closed,
-      // add the the bracket phrase that will eventually be added the the last node
-      if (bracketCount > 0) {
-        bracketPhrase += foundNode + phrase;
-      } else {
-        ast.set(foundNode, phrase);
-        lastNode = foundNode;
-      }
-      bracketCount += (phrase.match(/\(/g) || []).length;
-      bracketCount -= (phrase.match(/\)/g) || []).length;
-      if (bracketCount <= 0 && lastBracketCount > 0) {
-        // The phrase brackets is complete
-        // If it is a FROM phrase lets make a branch node
-        // If it is anything else lets make a leaf node
-        if (lastNode === 'FROM') {
-          ast.set(lastNode, this.sqlToAST(bracketPhrase));
-        } else {
-          const p = (ast.get(lastNode) as string).concat(bracketPhrase);
-          ast.set(lastNode, p);
-        }
-      }
-      lastBracketCount = bracketCount;
-    }
-    return ast;
+    return clausesToSql(filteredAST);
   }
 
   private applyFiltersToAST(ast: Map<string, Clause>, whereClause: string): Map<string, Clause> {
@@ -87,19 +50,6 @@ export class AdHocFilter {
     return ast.set('FROM', fromAST);
   }
 
-  private clausesToSql(ast: Map<string, Clause>): string {
-    let r = '';
-    ast.forEach((c: Clause, key: string) => {
-      if (typeof c === 'string') {
-        r += `${key} ${c.trim()} `;
-      } else if (c !== null) {
-        r += `${key} (${this.clausesToSql(c)} `;
-      }
-    });
-    // Remove all of the consecutive spaces to make things more readable when debugging
-    return r.trim().replace(/\s+/g, ' ');
-  }
-
   // Returns a table name found in the FROM phrase
   // FROM phrases might contain more than just the table name
   private getTableName(fromPhrase: string): string {
@@ -107,30 +57,6 @@ export class AdHocFilter {
       .trim()
       .split(' ')[0]
       .replace(/(;|\(|\))/g, '');
-  }
-
-  // Creates a statement with all the keywords to preserve the keyword order
-  private createStatement() {
-    let clauses = new Map<string, Clause>();
-    clauses.set('WITH', null);
-    clauses.set('SELECT', null);
-    clauses.set('DISTINCT', null);
-    clauses.set('FROM', null);
-    clauses.set('SAMPLE', null);
-    clauses.set('JOIN', null);
-    clauses.set('PREWHERE', null);
-    clauses.set('WHERE', null);
-    clauses.set('GROUP BY', null);
-    clauses.set('LIMIT BY', null);
-    clauses.set('HAVING', null);
-    clauses.set('LIMIT', null);
-    clauses.set('OFFSET', null);
-    clauses.set('UNION', null);
-    clauses.set('INTERSECT', null);
-    clauses.set('EXCEPT', null);
-    clauses.set('INTO OUTFILE', null);
-    clauses.set('FORMAT', null);
-    return clauses;
   }
 }
 
@@ -140,5 +66,3 @@ export type AdHocVariableFilter = {
   value: string;
   condition: string;
 };
-
-type Clause = string | Map<string, Clause> | null;
